popen: ignore setsid on macos
Due to os specifics we can't call setsid after
vfork on macos (vfork is not longer a part of
posix btw). Thus lets simply ignore it.

If there gonna be a strong need to drop
controlling terminal (we could use ioctl)
or setup new group we will enhance it.
But only if really needed.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>